Feat: moved Walk methods to Utils
